Accidentally noticed that mouseFlags in creature names is actually a string. Changed that in packet handling. Note that this requires database update too. Don't update to this rev yet ?
ALTER TABLE creature_names CHANGE mouseFlags info_str VARCHAR(500) NOT NULL


git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@2517 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/game/SpellEffects.cpp
@@ -348,7 +348,7 @@ void Spell::SpellEffectSchoolDMG(uint32 i) // dmg school
 		case 0xF4D5F002: //hunter - arcane shot
 			{
 				if(u_caster)
-					dmg += u_caster->GetRAP()*0.15f;
+					dmg += (uint32)(u_caster->GetRAP()*0.15f);
 			}break;
 		case 0xCBC738B8:	// Bloodthirst
 			{
